+++ lib/main.dart
@@ -5,6 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:intl/date_symbol_data_local.dart';
 import 'package:kakao_flutter_sdk_user/kakao_flutter_sdk_user.dart';
 import 'package:intl/intl.dart';
+import 'Ledger/view/tab_bar_view.dart';
 import 'home/view/ledger/tab_bar_view.dart';
 
 Future<void> main() async {
